separated the examples for attribute and dataset
